+++ qsdsan/sanunits/_heat_exchanging.py
@@ -16,13 +16,70 @@ Please refer to https://github.com/QSD-Group/QSDsan/blob/main/LICENSE.txt
 for license details.
 '''
 
-from biosteam.units import HXprocess, HXutility
+from biosteam.units import HXprocess as HXP, HXutility as HXU
+from biosteam.units.facilities import HeatExchangerNetwork as HXN
 from .. import SanUnit
 
-__all__ = ('HXprocess', 'HXutility',)
+__all__ = ('HeatExchangerNetwork', 'HXprocess', 'HXutility',)
+
+        
+class HeatExchangerNetwork(SanUnit, HXN):
+    '''
+    Similar to the :class:`biosteam.units.facilities.HeatExchangerNetwork`,
+    but also a subclass of :class:`qsdsan.SanUnit`
+
+    Examples
+    --------
+    >>> import qsdsan as qs
+    >>> from qsdsan import Stream, sanunits as su
+    >>> from qsdsan.utils import create_example_components
+    >>> qs.set_thermo(create_example_components())
+    >>> salt_water = Stream('salt_water', Water=2000, NaCl=50, units='kg/hr')
+    >>> methanol = Stream('methanol', Methanol=20, units='kg/hr')
+    >>> ethanol = Stream('ethanol', Ethanol=10, units='kg/hr')
+    >>> M1 = su.MixTank('M1', ins=(salt_water, 'recycled_brine', methanol, ethanol), init_with='Stream')
+    >>> P1 = su.Pump('P1', ins=M1-0, init_with='Stream')
+    >>> H1 = su.HXutility('H1', ins=P1-0, T=350, init_with='Stream')
+    >>> S1 = su.ComponentSplitter('S1', ins=H1-0, split_keys=('Methanol', 'Ethanol'), init_with='Stream')
+    >>> M2 = su.Mixer('M2', ins=(S1-0, S1-1), outs='alcohols', init_with='Stream')
+    >>> S2 = su.Splitter('S2', ins=S1-2, outs=(1-M1, 'waste_brine'), split=0.2, init_with='Stream')
+    >>> H2 = su.HXutility('H2', ins=S2.outs[1], T=280, init_with='Stream')
+    >>> HXN = su.HeatExchangerNetwork('HXN')
+    >>> sys = qs.System('sys', path=(M1, P1, H1, S1, M2, S2, H2), facilities=(HXN,))
+    >>> sys.simulate()
+    >>> # The actual utility usage is just 30% of the original one (i.e., without HXN)
+    >>> round(HXN.actual_heat_util_load/HXN.original_heat_util_load, 2)
+    0.28
+    >>> HXN.stream_life_cycles # doctest: +SKIP
+    [<StreamLifeCycle: Stream_0, cold
+     	life_cycle = [
+     		<LifeStage: <HXprocess: HX_0_1_hs>, H_in = 0 kJ, H_out = 3.15e+05 kJ>
+     		<LifeStage: <HXutility: Util_0_hs>, H_in = 3.15e+05 kJ, H_out = 4.4e+05 kJ>
+     	]>,
+     <StreamLifeCycle: Stream_1, hot
+     	life_cycle = [
+     		<LifeStage: <HXprocess: HX_0_1_hs>, H_in = 3.49e+05 kJ, H_out = 3.36e+04 kJ>
+     		<LifeStage: <HXutility: Util_1_cs>, H_in = 3.36e+04 kJ, H_out = -1.22e+05 kJ>
+     	]>]
+
+    See Also
+    --------
+    `biosteam.units.facilities.HeatExchangerNetwork <https://biosteam.readthedocs.io/en/latest/API/units/facilities/HeatExchangerNetwork.html>`_
+    '''
+    ticket_name = HXN.ticket_name
+    acceptable_energy_balance_error = HXN.acceptable_energy_balance_error
+    raise_energy_balance_error = HXN.raise_energy_balance_error
+    network_priority = HXN.network_priority
+    _N_ins = HXN._N_ins
+    _N_outs = HXN._N_outs
+    _N_heat_utilities = HXN._N_heat_utilities
+    _units= HXN._units
+    __init__ = HXN.__init__
+    _init_ins = HXN._init_ins
+    _init_outs = HXN._init_outs
 
 
-class HXprocess(SanUnit, HXprocess):
+class HXprocess(SanUnit, HXP):
     '''
     Similar to :class:`biosteam.units.HXprocess`,
     but can be initialized with :class:`qsdsan.SanStream` and :class:`qsdsan.WasteStream`.
@@ -31,6 +88,12 @@ class HXprocess(SanUnit, HXprocess):
     --------
     `biosteam.units.HXprocess <https://biosteam.readthedocs.io/en/latest/units/heat_exchange.html>`_
     '''
+
+    line = HXP.line
+    _graphics = HXP._graphics
+    _N_heat_utilities = HXP._N_heat_utilities
+    _N_ins = HXP._N_ins
+    _N_outs = HXP._N_outs
 
     def __init__(self, ID='', ins=None, outs=(), thermo=None,
                  init_with='Stream', F_BM_default=None,
@@ -83,7 +146,7 @@ class HXprocess(SanUnit, HXprocess):
         self.reset_source = True
 
 
-class HXutility(SanUnit, HXutility):
+class HXutility(SanUnit, HXU):
     '''
     Similar to :class:`biosteam.units.HXutility`,
     but can be initialized with :class:`qsdsan.SanStream` and :class:`qsdsan.WasteStream`.
@@ -92,6 +155,9 @@ class HXutility(SanUnit, HXutility):
     --------
     `biosteam.units.HXutility <https://biosteam.readthedocs.io/en/latest/units/heat_exchange.html>`_
     '''
+
+    line = HXU.line
+    _graphics = HXU._graphics
 
     def __init__(self, ID='', ins=None, outs=(), thermo=None,
                  init_with='Stream', F_BM_default=None,